Refactoring to move bulk out of reports_controller_decorator
