+++ app/views/spree/admin/reports/orders_and_fulfillment.html.haml
@@ -1,4 +1,4 @@
-= form_for @search, :url => spree.orders_and_fulfillment_admin_reports_path do |f|
+= form_for @report.search, :url => spree.orders_and_fulfillment_admin_reports_path do |f|
   = render 'date_range_form', f: f
 
   .row
@@ -30,7 +30,7 @@
 %table#listing_orders.index
   %thead
     %tr{'data-hook' => "orders_header"}
-      - @header.each do |heading|
+      - @report.header.each do |heading|
         %th=heading
   %tbody
     - @table.each do |row|
